+++ src/layout/dagre/src/order/init-order.ts
@@ -43,14 +43,14 @@ const initOrder = (g: Graph) => {
     return g.node(n).fixorder !== undefined;
   });
   const fixOrderNodes = beforeSort.sort((a, b) => (g.node(a).fixorder as number) - (g.node(b).fixorder as number));
-  fixOrderNodes.forEach((n) => {
+  fixOrderNodes?.forEach((n) => {
     if (!isNaN(g.node(n).rank as number)) {
       layers[g.node(n).rank as number].push(n);
     }
     visited[n] = true;
   });
 
-  orderedVs.forEach(dfs);
+  orderedVs?.forEach(dfs);
 
   return layers;
 };